+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Not release yet:
+* [Actions triggered when fields are changed](https://github.com/famiprog/redmine_goodies/pull/11)
+
 ## v1.1.1
 * [Collapsible section button available on every html editor](https://github.com/famiprog/redmine_goodies/pull/9)